+++ Changelog.md
@@ -3,3 +3,4 @@
 - Re-did the entire codebase and implemented hulu API to scrap data.
 - Proxy support added.
 - Fix for #22
+- Added Verbose Logging.